+++ contracts/finality/src/lib.rs
@@ -12,6 +12,7 @@ pub mod msg;
 pub mod queries;
 pub mod state;
 pub mod utils;
+pub mod validation;
 
 #[cfg_attr(not(feature = "library"), entry_point)]
 pub fn instantiate(